Cards and UI
Created the card base class, there is a lot of work that needs to happen 
here, right now we are just drawing some cards to test the scaling.

Also implemented the framework for the ui for mana and meditation 
crystals. These need the art fully flushed out and the images created, 
but then they will be ready to roll.