Art and GUI.
Moved a few GUI elements, did some work on the game board.

Added 1 - 10 for the meditation counter, so it can function.

Reworked the way that cards will work. There is now a library inside of 
cards.py that will match the cardlist.md inside the assets folder.

Added copyright text to each file.