First gameplay implementation test
The player can now draw cards by pressing space.

The only cards in the players inventory are mana cards, the player has 
one of each and enough MP to buy each of them.

They successfully add mana to the players mana counter and are then 
removed, not yet added to the graveyard.