Started work on playing cards - NOT WORKING - About to refactor all.
Everything is about to recieve a rework, a playerManager, cardManager, 
gameStateManager and imageManager will all be built.

The player is currently doing too much, I am not a fan of this. Working 
on planning out a series of managers that the player interacts with, 
they do all the heavy lifting.

This will be unfunctional for a little while.